434_analytics_application_engineering/Week_5/requirements.txt
Flask==2.0.1
google-cloud-storage==1.42.0
google-cloud-bigquery==2.25.1
gunicorn==20.1.0
db-dtypes == 1.1.1
pandas == 1.4.4